change from int to float

+++ bot/bot_utils.py
@@ -16,7 +16,7 @@ from pytz import timezone
 # =====================================
 
 
-def extract_number(user_input) -> Optional[int]:
+def extract_number(user_input) -> Optional[float]:
     """
     gets the amount of money from a string
     :param user_input: string with digits in it
@@ -26,7 +26,7 @@ def extract_number(user_input) -> Optional[int]:
     number = amount_pattern.search(user_input)
     if not number:
         return None
-    return int(number[0])
+    return float(number[0])
 
 
 def get_message_date(message: telegram.message.Message) -> datetime:
